// packages/oss-console/src/components/Breadcrumbs/types.ts
export interface Breadcrumb {
  /**
   * Unique to match the breadcrumb to the registry.
   * Also used for matching the url path segment.
   * Can be namespaced with a colon.
   */
  id: string;
  /**
   * The label to display for the breadcrumb.
   */
  label: string;
  /**
   * The default value of the breadcrumb button if not just setting from the URL.
   * This can be a string or a function that returns a string at runtime.
   */
  defaultValue: string | BreadcrumbCustomDefaultValue;
  /**
   * A function used for fetching the value of a breadcrumb at runtime.
   * Overides the behavior of the defaultValue if present.
   */
  asyncValue?: BreadcrumbAsyncValue;
  /**
   * A function to descide if the breadcrumb is visible on the page or not.
   * This can be used to hide breadcrumbs that are not relevant to the current page.
   */
  validator: BreadcrumbValidator;
  /**
   * A function used for fetching the popover data of a breadcrumb at runtime.
   * This is used for the popover list of items.
   */
  asyncData?: BreadcrumbAsyncPopOverData;
  /**
   * The title shown above the list within the popover.
   */
  popoverTitle?: string;
  /**
   * A react component to render instead of the default button and popover.
   * It recieved the same props as the default button and popover.
   */
  customComponent?: React.FC<any>;
  /**
   * The link that will used when a user clicks on the breadcrumb value text.
   * An empty string disables the link.
   */
  selfLink: BreadcrumbEntitySelfLink;
  /**
   * A function used for fetching the self link of a breadcrumb at runtime.
   */
  asyncSelfLink?: BreadcrumbAsyncViewAllLink;
  /**
   * The link to view all of a given collection of breadcrumb entities.
   */
  viewAllLink: BreadcrumbEntityViewAllLink;
  /**
   * A function used for fetching the view all link of a breadcrumb at runtime.
   */
  asyncViewAllLink?: BreadcrumbAsyncViewAllLink;
  /**
   * Show on all pages
   */
  required: boolean;
}
// ...
export interface BreadcrumbFormControlInterface extends Breadcrumb {
  value: string;
  key: string;
  projectId: string;
  domainId: string;
}

/**
 * The props used for rendering the breadcrumb list UI components.
 * Contains extra props for internal React rendering.
 */
export interface BreadcrumbFormControlInterfaceUI extends BreadcrumbFormControlInterface {
  children?: any;
  variant?: 'title' | 'inline';
}